+++ test/pipeline/pipeline-text-to-image.js
